+++ comtech-ecommerce/front-webadmin-reactjs/src/pages/Tag.jsx
@@ -7,13 +7,16 @@ import * as TagService from '../services/tagService';
 import { Dialog, DialogContent, DialogActions } from '@mui/material';
 import { useSelector } from 'react-redux';
 import { toast } from 'react-toastify';
-import { formatTimestamp } from '../utils/utils';
+import { decodeJWT, formatTimestamp } from '../utils/utils';
 import UpsertTag from '../components/Tag/UpsertTag';
 import OrderByBtn from '../components/OrderByBtn/OrderByBtn';
 
 export default function Tag() {
 
-  const authUser = useSelector(state => state.auth.user);
+  const authUser = useSelector(state => state.auth.user)
+  const authToken = useSelector(state => state.auth.accessToken)
+  const userRole = authToken ? decodeJWT(authToken).role : ''
+  
   const [loading, setLoading] = useState(false);
   const [onSubmit, setOnSubmit] = useState(false);
   const [refresh, setRefresh] = useState(0);
@@ -207,38 +210,44 @@ export default function Tag() {
     return result;
   }
   const handleDeleteTags = async () => {
-    setOnSubmit(true);
-    try {
-      const tagsId = deleteTagsId.map(i => parseInt(i));
-      const requestData = {
-        userId: authUser.id,
-        tagsId: deleteType === 'multiple' ? tagsId : [parseInt(selectDeleteTag)]
+    if(userRole === 'ADMIN') {
+      setOnSubmit(true);
+      try {
+        const tagsId = deleteTagsId.map(i => parseInt(i));
+        const requestData = {
+          userId: authUser.id,
+          tagsId: deleteType === 'multiple' ? tagsId : [parseInt(selectDeleteTag)]
+        }
+
+        //console.log(requestData)
+
+        await TagService.deleteTags(requestData)
+          .then(res => {
+            console.log(`Delete tags is successfully! : ${res.RESULT_DATA}`);
+            handleRefreshData();
+            setOnSubmit(false);
+            setConfirmDeletesDialog(false);
+            setSelectDeleteTag(null);
+            setDeleteTagsId([]);
+            handleResetToCreate();
+            toast.success(`Delete tags is successfully!`);
+          })
+          .catch(error => {
+            throw new Error(`Delete tags failed due to: ${error.response.data}`);
+          })
       }
-
-      //console.log(requestData)
-
-      await TagService.deleteTags(requestData)
-        .then(res => {
-          console.log(`Delete tags is successfully! : ${res.RESULT_DATA}`);
-          handleRefreshData();
-          setOnSubmit(false);
-          setConfirmDeletesDialog(false);
-          setSelectDeleteTag(null);
-          setDeleteTagsId([]);
-          handleResetToCreate();
-          toast.success(`Delete tags is successfully!`);
-        })
-        .catch(error => {
-          throw new Error(`Delete tags failed due to: ${error.response.data}`);
-        })
+      catch(error) {
+        console.log(error);
+        toast.error(`${error}`);
+      }
+      finally {
+        setOnSubmit(false);
+      }
     }
-    catch(error) {
-      console.log(error);
-      toast.error(error);
+    else {
+      toast.error(`You are in "Guest" mode, this action is not authorize.`)
     }
-    finally {
-      setOnSubmit(false);
-    }
+    
   }
 
   if(loading) return <div>กำลังโหลด...</div>
@@ -252,7 +261,7 @@ export default function Tag() {
 
       <div className="row">
         <div className="col-12 mb-3">
-          <div className="utils-head-table">
+          <div className="d-flex justify-content-end align-items-center">
             <button 
               className='btn my-btn purple-btn big-btn' 
               type="button"
